changed lidar quality metrics
