+++ map-structure/vi-map/include/vi-map/landmark-quality-metrics.h
@@ -1,15 +1,16 @@
 #ifndef VI_MAP_LANDMARK_QUALITY_METRICS_H_
 #define VI_MAP_LANDMARK_QUALITY_METRICS_H_
 
+#include <vector>
+
 #include <Eigen/Core>
 #include <aslam/common/memory.h>
-#include <vector>
 
 namespace vi_map {
 class VIMap;
 class Landmark;
-}  // namespace vi_map
-
+}
+   
 namespace vi_map {
 
 struct LandmarkWellConstrainedSettings {
@@ -19,28 +20,13 @@ struct LandmarkWellConstrainedSettings {
   double max_distance_to_closest_observer;
   /// Maximum distance from closest observer for a LiDAR landmark to be well
   /// constrained [m].
-  double max_distance_to_closest_observer_lidar;
-  /// Minimum distance from closest observer for a landmark to be well
-  /// constrained [m].
   double min_distance_to_closest_observer;
   /// Minimum distance from closest observer for a LiDAR landmark to be well
   /// constrained [m].
-  double min_distance_to_closest_observer_lidar;
-  /// Minimum angle disparity of observers for a landmark to be well constrained
-  /// [deg].
   double min_observation_angle_deg;
   /// Minimum angle disparity of observers for a LiDAR landmark to be well
   /// constrained [deg].
-  double min_observation_angle_deg_lidar;
-  /// Minimum number of observers for a landmark to be well constrained.
   size_t min_observers;
-  /// Minimum number of observers for a LiDAR landmark to be well constrained.
-  size_t min_observers_lidar;
-  /// Maximum distance between landmark measurements of a LiDAR [m].
-  double max_position_deviation_lidar;
-  /// Maximal ratio between the distance between LiDAR landmark measurements and
-  /// the distance from the observer to the closest measurement..
-  double max_position_uncertainty_lidar;
 
   LandmarkWellConstrainedSettings();
 };
@@ -61,16 +47,6 @@ inline bool isLandmarkWellConstrained(
 bool isLandmarkWellConstrained(
     const vi_map::VIMap& map, const vi_map::Landmark& landmark,
     bool re_evaluate_quality);
-
-bool isLidarLandmarkWellConstrained(
-    const vi_map::VIMap& map, const vi_map::Landmark& landmark,
-    bool re_evaluate_quality);
-
-// Includes an additional check for 3D LiDAR data
-bool isLidarLandmarkWellConstrained(
-    const vi_map::VIMap& map, const vi_map::Landmark& landmark,
-    bool re_evaluate_quality, double min_distance_to_lidar,
-    double position_uncertainty);
 
 }  // namespace vi_map
 #include "./landmark-quality-metrics-inl.h"
